pyproject.toml
[project]
name = "wavereader"
version = "0.1.0"
requires-python = ">=3.14.0"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "uvicorn[standard]>=0.31.0",
    "requests>=2.32.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.0.0",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.14.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"main.py" = ["E402"]
"data_gen/*" = ["E501"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
